disaster-analysis/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# AI/ML
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
ftfy>=6.1.0
regex>=2023.0.0

# Image/Video processing
opencv-python>=4.8.0
Pillow>=10.0.0

# Utilities
pydantic>=2.0.0
python-dotenv>=1.0.0
